[Refactor] util 통합
