--- app/src/main/java/com/and04/naturealbum/utils/ImageUtils.kt
@@ -1,28 +1,0 @@
-package com.and04.naturealbum.utils
-
-import android.content.Context
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
-import android.net.Uri
-import android.util.Base64
-import java.io.ByteArrayOutputStream
-
-fun Context.getBase64FromUri(uriString: String): String {
-    return try {
-        val uri = Uri.parse(uriString)
-
-        val options = BitmapFactory.Options().apply { inSampleSize = 16 } // 이미지 크기 축소
-        val bitmap = this.contentResolver.openInputStream(uri)?.use {
-            BitmapFactory.decodeStream(it, null, options)
-        } ?: return ""
-
-        val byteArrayOutputStream = ByteArrayOutputStream()
-        bitmap.compress(Bitmap.CompressFormat.JPEG, 100, byteArrayOutputStream) // 압축 품질
-        val byteArray = byteArrayOutputStream.toByteArray()
-
-        Base64.encodeToString(byteArray, Base64.DEFAULT)
-    } catch (e: Exception) {
-        e.printStackTrace()
-        ""
-    }
-}
