+++ gradle/libs.versions.toml
@@ -35,6 +35,7 @@ firebaseStorageKtx = "21.0.1"
 tensorflowLite = "2.12.0"
 tensorflowLiteTaskVision = "0.4.0"
 retrofit = "2.6.4"
+benchmarkCommon = "1.3.3"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
@@ -89,6 +90,7 @@ firebase-storage-ktx = { group = "com.google.firebase", name = "firebase-storage
 tensorflow-lite = { module = "org.tensorflow:tensorflow-lite", version.ref = "tensorflowLite" }
 tensorflow-lite-task-vision = { module = "org.tensorflow:tensorflow-lite-task-vision", version.ref = "tensorflowLiteTaskVision" }
 retrofit = { module = "com.squareup.retrofit2:retrofit", version.ref = "retrofit" }
+androidx-benchmark-common = { group = "androidx.benchmark", name = "benchmark-common", version.ref = "benchmarkCommon" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
